Reduce needless recompiles
Build script calls skeptic which emits information asking cargo to
monitor content of README.md.skt.md for changes. If changes to that file
are detected cargo should rebuild this crate.  In case of this project
the files does not exist. This causes cargo to trigger a rebuild every
time in hope that it will get created by the build script.

The workaround is to create the file that skeptic asks to monitor which
stops cargo from triggering the rebuild.

For details consult this issues:

* budziq/rust-skeptic#25
* rust-lang/cargo#4213