Update to difference v2.0
This updates the difference dependency to version 2, which introduces a
slightly rewritten difference algorithm and was failing tests for some
crates (including this one). Hence the major version bump.

The failing test in this case was just some white space being prepended
to the output (presumably because difference added an empty string to the
beginning of its output before), so that looks more like a fix to me.