feat(output): Generic predicate support
This accepts a message with it.  This should hit the 90% case of a
`satisfies_ok` (or whatever it'd be called).

I'm also assuming that it'll be a best practice to document the custom
predicates, so its acceptable to force it on everyone.

If a `satisfies_ok` is found to be needed, I'm assuming its because the
user wants to tie into existing machinery that has error reporting.  This
means we'll probably need to accept an `Fn` that `Box`es the error to
preserve it.

Fixes #55