refactor: Migrate to failure
This takes the use of `Context` to a bit of an extreme, using a single
error type and propogating any additional information along as
individual pieces of `Context.

Example:
```
command="wc" "README.md"
with: stderr='''wc: README.md: No such file or directory
'''
with: stdout=''''''
with: expected to succeed
with: CLI Assertion Error: Status mismatch.
```

BREAKING CHANGE: The error type of `Assert` is now `failure::Error`.