Add optional access to Command's stdin handle
Add access to Command's stdin handle

Add InputPredicate type for stdin

This trait allows for the computation of the contents to be written to
stdin using the `stdin` method of the `Assert` type. A series of
`From<T>` implementations are added for the trait so that there is no
functionality change to the `stdin` method.

By changing to the trait, a "handle" is provided for a user to
access stdin during a test. There is no functionality change to the
public API, but the `stdin` method could be used in a chain to write
content to stdin over time. This can be useful when testing the
functionality of a CLI application with streaming input.

Add `stdin_contents` field to debug implementation

Change lifetimes for trait implementation

The 'static lifetime requirement is changed to 'a for the From<&str>
implementation of the InputPredicate type. This makes its usage a little
more flexible.

Add example with a delay

An example of composing stdin content over time as a stream to the CLI
application under test is added to the `stdin` method.

Fix tests

Add documentation to `InputPredicate` trait

Remove excessive whitespace

Add some documentation to explain examples

Add `From` implementation

A `From<&[u8]>` implementation was missing for the `InputPredicate`
trait to avoid a regression.

Add export of `InputPredicate` to public API

This is needed so users can define types that implement the trait in
their tests.

Change `From` trait bounds to be generic

After some usage with creating another test, types that implemented the
`InputPredicate` trait also needed to implement the `From` trait for the
`Box<InputPredicate>` to work. So, the built-in `From` trait to create
a boxed `InputPredicate` is changed to be more generic, so (maybe?) the
conversion trait does not need to be implement for all types.

Refactor formatting for consistency

Add example of using type

An example is added for the `stdin` method that demonstrates how to
define an `InputPredicate` implementation using a unit struct instead of
a closure. This improves code re-use and arguably readability/fluency.

Add `example` constructor

The `example` constructor is added to the `Assert` struct as a helper
method to run an example from the crate instead of having to use the
`command` constructor. This is similar to the `cargo_binary` but uses
the `--example` option instead of the `--bin` option.

Refactor formatting with rustfmt

Fix comment grammer

Refactor formatting

Add `stdin_contents` field to debug implementation

Fix tests

Add documentation to `InputPredicate` trait

Remove excessive whitespace

Add export of `InputPredicate` to public API

This is needed so users can define types that implement the trait in
their tests.

Refactor formatting for consistency

Add `example` constructor

The `example` constructor is added to the `Assert` struct as a helper
method to run an example from the crate instead of having to use the
`command` constructor. This is similar to the `cargo_binary` but uses
the `--example` option instead of the `--bin` option.

Refactor formatting with rustfmt

Fix comment grammer

Refactor formatting