fix(cargo): Avoid `cargo run`
This switches us from using `cargo run` to `cargo build`, reading where
the binary is placed, and callin that instead.

Fixes #95 because the user changing the `CWD` doesn't impact `cargo
build`.

Fixes #79 because there is no `cargo` output when capturing the user's
stdout/stderr.

Fixes #51 because the user changing the environment doesn't impact
`cargo build`.

This is a step towards working around #100 because we can allow a user
to cache the result of `cargo build`.